+++ events/name_onInput.js
@@ -1,9 +1,9 @@
 // Enables/disables create folder submit button
-const folderName_onInput = (e) => {
+const name_onInput = (e) => {
   const input = e.target;
-  const submitButton = document.querySelector(
-    '.form-create-folder button[type="submit"]'
-  );
+  const form = input.form;
+  const submitButton = form.querySelector('button[type="submit"]');
+
   if (input.value.trim().length === 0) {
     submitButton.setAttribute("disabled", "");
   } else {